llm calls moved inside file
